constexpr not supported by Visual C++ 2013 (even after update 2) so switched to dumber approach
